+++ website_sale_checkout_store/controllers/main.py
@@ -52,20 +52,23 @@ class WebsiteSaleExtended(WebsiteSale):
         else:
             return super(WebsiteSaleExtended, self).payment_get_status(sale_order_id, **post)
 
-    def _get_mandatory_fields(self):
+    def _get_mandatory_fields(self, shipping=None):
         order = request.website.sale_get_order()
         if not order.buy_way or 'nobill' not in order.buy_way and 'noship' not in order.buy_way:
-            return ["name", "phone", "email", "street", "city", "country_id"]
+            fields = ["name", "phone", "street", "city", "country_id"]
         elif 'noship' in order.buy_way:
             if 'nobill' in order.buy_way:
-                return ["name", "phone", "email"]
+                fields = ["name", "phone"]
             else:
-                return ["name", "phone", "email", "country_id"]
+                fields = ["name", "phone", "country_id"]
         else:
-            return ["name", "phone", "email", "street", "city"]
+            fields = ["name", "phone", "street", "city"]
+        if not shipping:
+            fields.append('email')
+        return fields
 
     def _get_mandatory_billing_fields(self):
         return self._get_mandatory_fields()
 
     def _get_mandatory_shipping_fields(self):
-        return self._get_mandatory_fields()
+        return self._get_mandatory_fields(shipping=True)